% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(shiny.maxRequestSize = 1024 * 1024^2, ...)
}
\arguments{
\item{shiny.maxRequestSize}{This is a number which specifies the maximum web request size, which 
serves as a size limit for file uploads. If unset, the maximum request size defaults to 1024 MiB.}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
